Move mesh_socket.hpp to experimental
