+++ tests/ttnn/unit_tests/gtests/multiprocess/test_send_recv_pipeline.cpp
@@ -11,7 +11,7 @@
 #include "ttnn/distributed/distributed_tensor.hpp"
 
 #include "tt_metal/multihost/fabric_tests/multihost_fabric_fixtures.hpp"
-#include <tt-metalium/mesh_socket.hpp>
+#include <tt-metalium/experimental/sockets/mesh_socket.hpp>
 #include <tt-metalium/distributed_context.hpp>
 #include "tt_metal/fabric/physical_system_descriptor.hpp"
