+++ tt_metal/CMakeLists.txt
@@ -81,7 +81,7 @@ set(TT_METAL_PUBLIC_API
     api/tt-metalium/mesh_event.hpp
     api/tt-metalium/experimental/fabric/mesh_graph.hpp
     api/tt-metalium/experimental/fabric/mesh_graph_descriptor.hpp
-    api/tt-metalium/mesh_socket.hpp
+    api/tt-metalium/experimental/sockets/mesh_socket.hpp
     api/tt-metalium/mesh_trace_id.hpp
     api/tt-metalium/mesh_workload.hpp
     api/tt-metalium/metal_soc_descriptor.h
