+++ tt_metal/distributed/mesh_socket.cpp
@@ -224,7 +224,7 @@ void MeshSocket::connect_with_peer(const std::shared_ptr<multihost::DistributedC
 
     std::vector<Rank> sender_ranks = get_ranks_for_mesh_id(config_.sender_mesh_id.value(), rank_translation_table_);
     std::vector<Rank> recv_ranks = get_ranks_for_mesh_id(config_.receiver_mesh_id.value(), rank_translation_table_);
-    // Barrier across all sender and receiver ranks. This ensures that the downstream workloads using these socket
+    // Barrier across all sender and receiver ranks. This ensures that the downstream workloads using this socket
     // will start after the socket is initialized across all hosts.
     execute_with_timeout([&]() { barrier_across_send_recv_ranks(sender_ranks, recv_ranks, context); });
 }
