+++ ttnn/cpp/ttnn-nanobind/mesh_socket.cpp
@@ -13,7 +13,7 @@
 #include <nanobind/stl/optional.h>
 #include <nanobind/stl/shared_ptr.h>
 
-#include <tt-metalium/mesh_socket.hpp>
+#include <tt-metalium/experimental/sockets/mesh_socket.hpp>
 
 namespace ttnn::mesh_socket {
