+++ ttnn/cpp/ttnn/operations/experimental/ccl/send_recv_async/send_recv_utils.hpp
@@ -9,7 +9,7 @@
 #include <unordered_set>
 #include <enchantum/enchantum.hpp>
 
-#include <tt-metalium/mesh_socket.hpp>
+#include <tt-metalium/experimental/sockets/mesh_socket.hpp>
 #include "ttnn/tensor/tensor.hpp"
 
 namespace ttnn::send_recv_utils {
